Fix numpy float64 leaking into raw SQL params; document Transaction Pooler failure
smart_money_pipeline.py computed pct_price_chg/action/dlv_pct/pct_oi_chg via
round() on numpy scalars from pd.to_numeric() without casting to float first.
psycopg2 can't adapt numpy types, so it fell back to inserting numpy's repr
text raw into the query, producing "np.float64(...)" in the SQL and either a
"schema np does not exist" error or a broken connection reported as an SSL
error, depending on symbol/timing — same root cause behind both error
messages seen in production. Wrapped each in float() before round().

Also confirmed and reverted the port 6543 (Transaction Pooler) attempt from
the previous commit: it's actively broken for querying on this Supabase
project (connect succeeds, first query then fails with SSL connection
closed, 100% reproducible). Session Pooler (5432) is the working
configuration; db.py's comments now document this so it isn't re-attempted
blindly.

+++ backend/data_ingestion/smart_money_pipeline.py
@@ -90,9 +90,9 @@ def _fetch_one_day(symbol: str, dt: date) -> dict | None:
                 row["trade_qty"]     = float(tqty)  if pd.notna(tqty)  else None
                 row["tot_trade"]     = float(ttrd)  if pd.notna(ttrd)  else None
                 if pd.notna(close) and pd.notna(prev) and prev > 0:
-                    row["pct_price_chg"] = round((close - prev) / prev * 100, 2)
+                    row["pct_price_chg"] = round(float((close - prev) / prev * 100), 2)
                 if pd.notna(tqty) and pd.notna(ttrd) and ttrd > 0:
-                    row["action"] = round(tqty / ttrd, 1)
+                    row["action"] = round(float(tqty / ttrd), 1)
                 got_any = True
     except Exception:
         pass
@@ -110,7 +110,7 @@ def _fetch_one_day(symbol: str, dt: date) -> dict | None:
                 qty = pd.to_numeric(m["qty_traded"].iloc[0], errors="coerce")
                 dlv = pd.to_numeric(m["dlv_qty"].iloc[0],   errors="coerce")
                 if pd.notna(qty) and pd.notna(dlv) and qty > 0:
-                    row["dlv_pct"] = round(dlv / qty * 100, 2)
+                    row["dlv_pct"] = round(float(dlv / qty * 100), 2)
                 got_any = True
     except Exception:
         pass
@@ -133,7 +133,7 @@ def _fetch_one_day(symbol: str, dt: date) -> dict | None:
                     row["oi_change"]  = float(oichg) if pd.notna(oichg) else None
                     prev_oi = oi - (oichg if pd.notna(oichg) else 0)
                     if prev_oi != 0:
-                        row["pct_oi_chg"] = round((oichg / prev_oi) * 100, 2)
+                        row["pct_oi_chg"] = round(float((oichg / prev_oi) * 100), 2)
                 got_any = True
     except Exception:
         pass
